+++ lib/api/SetAttribute.ts
@@ -73,12 +73,12 @@ const createImpressionForAttributes = async (
   // Retrieve base properties for the event
   const properties = getEventsBaseProperties(
     serviceContainer.getSettingsService(),
-    EventEnum.VWO_SYNC_VISITOR_PROP,
+    EventEnum.SYNC_VISITOR_PROP,
     encodeURIComponent(context.getUserAgent()),
     context.getIpAddress(),
   );
   // Construct payload data for multiple attributes
-  const payload = getAttributePayloadData(serviceContainer, EventEnum.VWO_SYNC_VISITOR_PROP, attributes, context);
+  const payload = getAttributePayloadData(serviceContainer, EventEnum.SYNC_VISITOR_PROP, attributes, context);
 
   if (serviceContainer.getBatchEventsQueue()) {
     serviceContainer.getBatchEventsQueue().enqueue(payload);