+++ lib/utils/LogMessageUtil.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 import { isFunction } from '../utils/DataTypeUtil';
+import { Constants } from '../constants';
 
 const nargs = /\{([0-9a-zA-Z_]+)\}/g;
 
@@ -25,6 +26,11 @@ const nargs = /\{([0-9a-zA-Z_]+)\}/g;
  * @returns {string} The constructed message with all placeholders replaced by their corresponding values from the data object.
  */
 export function buildMessage(template: string = '', data: Record<string, any> = {}): string {
+  const payload = {
+    brand: Constants.BRAND_DISPLAY_NAME,
+    logPrefix: Constants.LOG_PREFIX,
+    ...data,
+  };
   try {
     return template.replace(nargs, (match, key, index) => {
       // Check for escaped placeholders
@@ -33,7 +39,7 @@ export function buildMessage(template: string = '', data: Record<string, any> = 
       }
 
       // Retrieve the value from the data object
-      const value = data[key];
+      const value = payload[key];
 
       // If the key does not exist or the value is null/undefined, return an empty string
       if (value === undefined || value === null) {