+++ lib/utils/XMLUtil.ts
@@ -113,7 +113,7 @@ function sendRequest(method, options, logManager: LogManager) {
         requestModel.setLastError(error);
         setTimeout(executeRequest, delay);
       } else {
-        if (!String(networkOptions.path).includes(EventEnum.VWO_DEBUGGER_EVENT)) {
+        if (!String(networkOptions.path).includes(EventEnum.DEBUGGER_EVENT)) {
           logManager.errorLog(
             'NETWORK_CALL_FAILURE_AFTER_MAX_RETRIES',
             {