+++ test/unit/packages/segmentation-evaluator/Segmentation.test.ts
@@ -34,7 +34,7 @@ const createMockServiceContainer = (): ServiceContainer => {
     getSettings: () => ({}),
     getSettingsService: () => ({}),
     getHooksService: () => ({}),
-    getVWOOptions: () => ({}),
+    getWingifyOptions: () => ({}),
     getBatchEventsQueue: () => ({}),
     getSegmentationManager: () => ({}),
     getBaseUrl: () => '',